requirements.txt
streamlit 
PyPDF2
python-dotenv
langchain
sentence-transformers
faiss-cpu
dotenv
requests